--with-pybank option removed.
Paygen must be run against Python bank, as all the
components involved uses real databases on the hosting
system, and so must do the bank.

Fakebank is not able to fit in this picture, as it
uses a in-memory (mock) database, which starts "fresh"
upon every invocation and whose data cannot be consistent
with data that exists into the ordinary databases used
by other components.