/track/transaction CMD stores WTID as trait.
This is needed because the real bank does not provide
a "bank check" CMD as the fakebank does, making it impossible
to get any WTID while running the paygen against the real
bank.  This way, whenever we use the /track/transaction CMD
we get the WTID stored by it, and usable by future CMDs.