/proposal API disambiguation.
The order object gives the client TWO ways to specify
the merchant instance: one is at the object top-level,
and the other one is within the "merchant" sub-object.

This change makes sure that the top-level instance
indication and the one internal to the merchant sub-object
refer always to the same instance.

Note: when the top-level indication is missing, then
the instance is set to "default", and the merchant sub-
object must likewise specify a "default" instance.